Initial commit of some hgpstools scripts
The code in the scripts my_ip.pl, serialdaemon_gps.pl, and udplisten.pl
is almost identical to that used successfully in a previous experiment
using a Raspberry Pi (the changes were mostly: getopts and pod2usage
were added to all scripts, and the minimum Perl version was reduced from
v5.14 to v5.10). All of the POD documentation, README and filter_ts.pl
were added recently.