Use "umask" option of (patched) Daemon::Daemonize
Replace the logic which chmods the files after daemonizing with the
"umask" option that was added in our patched version of
Daemon::Daemonize. This commit also pulls in the new version of our
fork which includes the version number update to 0.0052_01, so that we
can properly reference it in our "use" statement. Also added some
information to the README about the submodule.