Updated serial port "abort" (signal) handling
In SerialPort, merged the "abort" and "aborted" flags; there is now only
"abort" (with a getter named "aborted"); this must be cleared manually
instead of being reset automatically by "read". Also, "read" now returns
immediately when "abort" is set.

The reason for this change is that in ngserlog.pl, the "on start" and
"on connect" handlers did not have a good way of making use of the
signal handlers set by the main code. Updated ngserlog.pl accordingly.

Also made very minor doc and whitespace updates.