Adding the various logger scripts
This commit moves ngserlog_nmea.pl to the "serloggers" subdir and adds
the logger scripts for the Novatel SPAN-IGM-S1 (three loggers, one for
each of the USB1/2/3 ports, each of which we use for a different
purpose), the Mensor CPT6100 and the Vaisala HMT310. Note that these are
all still alpha or at best beta status, they need further testing and
documentation.