Added specific local::lib path to scripts (temp. fix!)
In order to get the daemons up and running in the test environment, I'm
hard-coding the local::lib and "use lib" paths for now. There are much
better solutions, like setting up the libraries in more general
locations and then pointing Perl at them, but this is the quickest way
for now. This should be updated later.